Khushal: Fixed a visualization query bug

+++ app/models/visualization.rb
@@ -1,25 +1,23 @@
 class Visualization
 
-  def self.getNoOfTests(sub_project_id,test_type)
+  def self.getNoOfTests(sub_project_id, test_type)
 
     meta_data = SubProject.find(sub_project_id).test_metadatum.find_all_by_test_category(test_type)
 
-    meta_data.inject([]){ |result, metadata_record|
+    meta_data.inject([]) { |result, metadata_record|
       total_num_of_tests = 0
       total_run_time = 0
       metadata_record.test_suite_records.each do |test_suite_record|
         total_num_of_tests += test_suite_record.number_of_tests.to_i
-        total_run_time += test_suite_record.time_taken.to_i
+        total_run_time += test_suite_record.time_taken.to_f
       end
 
-      result << [total_num_of_tests,total_run_time]
+      result << [total_num_of_tests, total_run_time]
 
 
     }
 
   end
-
-
 
 
 end
